Minor build process updates with a new launcher script.

+++ install.sh
@@ -1,12 +1,17 @@
 #!/bin/bash
 
+# check .env is setup
+if [ ! -f '.env' ]; then
+    echo "docker-compose .env file is not found, please complete project setup before running install.sh"
+    exit
+fi
+
 # Install composer in to local user exec directory on OSX
 echo "Installing Composer via homebrew."
 brew install composer
 
 # Install package dependencies
-cd src
-composer install --no-interaction --no-ansi
+composer install --no-interaction --no-ansi -d ./src
 
 # and run docker-compose
 docker-compose up
